Add support for multiple datasets with Add Dataset
eCharts supports Chart's `dataset` as an array and indexing it
in a Series with `datasetIndex`.

This replaces the `Dataset` Chart config field with the
function `AddDataset`.  It also adds `WithDatasetIndex` as
an option for `SeriesOpts`

The `DatasetList` is JSON rendered as a single object, as
before.  But multiple Datasets are rendered as Array.

This is a breaking change:

```
// old
chart.Dataset = opts.Dataset{Source: test}
// new
chart.AddDataset(opts.Dataset{Source: test})
```